Create Maintenance

+++ app/src/main/java/br/com/lambdateam/mycar/model/maintenance/Maintenance.kt
@@ -2,9 +2,9 @@ package br.com.lambdateam.mycar.model.maintenance
 
 data class Maintenance(
     val id: Int?,
-    val km: Double?,
+    val km: Int?,
     val maintenanceDate: String?,
-    val nextKm: Double?,
+    val nextKm: Int?,
     val amount: Double?,
     val manufacturer: MaintenanceManufacturer?,
     val vehicle: MaintenanceVehicle?,
